Add Human Redux in Learning Material section
I'd love to include a git book called "Human Redux". 

I read through Human Redux and I found it a very good learning material for understanding the Redux concepts. It offers intuitive explanations with good examples and it shows the readers how to refactor our code to make it more readable and scalable.

The changes include
- add Human Redux under Redux's Concepts in the Learning Material section.